if the sidebar is already expanded and I open the user menu, when I close the user menu, it should collapse the whole sidebar.
I think that are multiple state defining the collapse/expand nature of the sidebar. When I click the user menu it expands sidebar and usermenu, and when i click it to close, it correctly closes both.

But, when the sidebar is already open and i click the user menu, and after i close th eu ser menu it does not close the sidebar. It should alwauys collapse the user menu and sidebar when user menu is collapsed.