I need to create 2 items on my menu.
I have a Recipe Generator, but now i want to create

Design System - a page where i will put all my design system elements and direction so i can reference and check it to later use.

Loading State - a page showing my loading feature that will be a loading state later on the development.

I want to create these 2 item menus so i can check them visually before implementing it on the workflow of the app.

Create also a task.md file for each of new the menu tab

Consult the files for specifications and templates to follow:
/home/user/studio/app_docs/01_design_system_01.html
/home/user/studio/app_docs/01_loading_feature.html

Be as consistent as possible with the refered documents above.