on the Chat Input (New Style)
I need the magnify glass as on the old style on the left with the separator | as it is right now and remove the other clickable items on the left. Magnify glass will not be clickable.

Use the same border effect as in the current old style for hover and focused states. with a thinner border