this is the expected visual.
Top should touch the header and be full width on mobile
"Top border" is the gradient progress bar.

Desktop should be a box with shadow without touching the top
Mobile be a full width without shadow box.

Be carefull because the image just represent expected behavior, and not theme colors and styles.